Initial Register form implementation
editied the register form and register page files, the register page displayes the form aswell as the blue background with the white circle as per the design. The register for current is not styled using css, however does interact with the api using the /register/ endpoint. Errors need to be added and a redirect after the user has successfully submitted a form.