Prototype register form complete
register form functions and is able to post a new user to the api. A general error is thrown when registration does not work, an error pop up is displayed if the passwords do not match.

the user is then redirected to the login page once the form submits and they register correctly