Created component for account details and initialized profile page
Included some small changes to fix errors on the prior 'account-info-segment' branch from sprint 3. The AccountDetails file is the main component and the GroupBar file is a subcomponent for the groups display. Made some small changes to the root file in order to get the links routing properly on the profile page.