Creating a media folder.
Altering fields on group, post and user to point the pictures and files to the media file.
We can now grab images using localhost/media/file_name/image_name :D big update.
In serializers added banner image only to groups, will need to add to image to post serializer when we are working on it.